beta
changes added
we added a proper logo
-SASS file  has been changed to support more responsive  featres proper
margins and colors
-logo has been aded its not exactly ideal but quickly aded for
dimensions and space limitations
-js files suffered a lot of changes ,  range has been increased to 250
miles to  accomodate a  few locations without a space, and  be able to
show  radious on montreall we also added a   zoom    event to stop  it
from being too close if only one location and we also  added  the
ability to  fit all locatons on statup as needed. most changes are
related to map  functions.

google keys for testing have been censored for  safety as all testing is
local atm please inster your own key for   further  items